Dialogs/WifiDialog: extract WifiDialog
